+++ docs/CHANGELOG.md
@@ -3,6 +3,11 @@
 
 ## [Unreleased]
 
+### Added
+
+- feat: breaks `readState` into `signed` and `unsigned` methods. The `signed` method will take a pre-signed request and re-use it for polling. The `unsigned` method will create a new request and sign it for each poll.
+  - This can be called directly in the `HttpAgent` or through the `Actor` interface, with a new `signed` polling strategy.
+
 ## [2.4.1] - 2025-04-10
 
 ### Changed